requirements.txt
Flask==1.1.2
google-cloud-datastore==1.15.0
google-cloud-storage==1.31.0
google-cloud-vision==1.0.0
